refactoring: separation of abstract report and derived report instances simplify InvoiceReport-parameters add generic threading support to printig and pdf export
