+++ src/main/java/kernbeisser/Windows/CashierShoppingMask/CashierShoppingMaskModel.java
@@ -3,17 +3,15 @@ package kernbeisser.Windows.CashierShoppingMask;
 import java.util.Collection;
 import java.util.function.Consumer;
 import javax.persistence.EntityManager;
-import javax.swing.ProgressMonitor;
 import kernbeisser.DBConnection.DBConnection;
 import kernbeisser.DBEntities.User;
+import kernbeisser.Enums.SaleSessionType;
 import kernbeisser.Enums.Setting;
-import kernbeisser.Exeptions.InvalidVATValueException;
-import kernbeisser.Reports.ReportManager;
+import kernbeisser.Reports.AccountingReport;
 import kernbeisser.Useful.Tools;
 import kernbeisser.Windows.LogIn.LogInModel;
 import kernbeisser.Windows.MVC.IModel;
 import lombok.Data;
-import net.sf.jasperreports.engine.JRException;
 
 @Data
 public class CashierShoppingMaskModel implements IModel<CashierShoppingMaskController> {
@@ -24,40 +22,28 @@ public class CashierShoppingMaskModel implements IModel<CashierShoppingMaskContr
   private boolean shoppingMaskOpened = false;
 
   void printTillRoll(Consumer<Boolean> resultConsumer) {
-    new Thread(
-            () -> {
-              ProgressMonitor pm =
-                  new ProgressMonitor(
-                      null, "Ladendienst wird gedruckt", "Initialiesiere Druckerservice...", 0, 2);
-
-              EntityManager em = DBConnection.getEntityManager();
-              long id =
-                  em.createQuery(
-                          "select max(p.id) from Purchase p where p.session.seller.id = :lid",
-                          Long.class)
-                      .setParameter("lid", LogInModel.getLoggedIn().getId())
-                      .getSingleResult();
-              long from = Setting.LAST_PRINTED_BON_NR.getLongValue();
-              if (from == id) {
-                resultConsumer.accept(false);
-                return;
-              }
-              Setting.LAST_PRINTED_BON_NR.changeValue(id);
-              try {
-                ReportManager.initAccountingReportPrint(from + 1, id);
-                pm.setProgress(1);
-                pm.setNote("Exportiere PDF..");
-                new ReportManager().exportPdf();
-              } catch (JRException | InvalidVATValueException e) {
-                Setting.LAST_PRINTED_BON_NR.changeValue(from);
-                resultConsumer.accept(false);
-                Tools.showUnexpectedErrorWarning(e);
-              }
-
-              pm.setNote("Fertig");
-              pm.close();
-              resultConsumer.accept(true);
-            })
-        .start();
-  }
+    EntityManager em = DBConnection.getEntityManager();
+    long id =
+        em.createQuery(
+                "select max(p.id) from Purchase p where p.session.seller.id = :lid and p.session.sessionType = :t",
+                Long.class)
+            .setParameter("lid", LogInModel.getLoggedIn().getId())
+            .setParameter("t", SaleSessionType.ASSISTED)
+            .getSingleResult();
+    long from = Setting.LAST_PRINTED_BON_NR.getLongValue();
+    if (from == id) {
+      resultConsumer.accept(false);
+      return;
+    }
+    Setting.LAST_PRINTED_BON_NR.changeValue(id);
+    AccountingReport accountingReportManager = new AccountingReport(from + 1, id);
+    accountingReportManager.exportPdf(
+        "Ladendienst wird gedruckt",
+        (e) -> {
+          Setting.LAST_PRINTED_BON_NR.changeValue(from);
+          resultConsumer.accept(false);
+          Tools.showUnexpectedErrorWarning(e);
+        });
+    resultConsumer.accept(true);
+  };
 }
