+++ src/main/java/kernbeisser/Windows/Tillroll/TillrollController.java
@@ -1,14 +1,10 @@
 package kernbeisser.Windows.Tillroll;
 
-import java.awt.print.PrinterAbortException;
 import kernbeisser.Enums.ExportTypes;
 import kernbeisser.Exeptions.IncorrectInput;
-import kernbeisser.Exeptions.InvalidVATValueException;
 import kernbeisser.Useful.Tools;
 import kernbeisser.Windows.MVC.Controller;
 import lombok.var;
-import net.sf.jasperreports.engine.JRException;
-import org.apache.commons.lang3.exception.ExceptionUtils;
 
 public class TillrollController extends Controller<TillrollView, TillrollModel> {
 
@@ -22,37 +18,34 @@ public class TillrollController extends Controller<TillrollView, TillrollModel> 
 
   public void exportTillroll(ExportTypes exportType, int days) {
     var view = getView();
+    model.exportTillroll(exportType, days, (e) -> consumePdfException(e, exportType));
+    view.back();
+  }
+
+  private void consumePdfException(Throwable e, ExportTypes exportType) {
     try {
-      model.exportTillroll(exportType, days);
-      view.back();
-    } catch (UnsupportedOperationException e) {
-      view.messageNotImplemented(exportType);
-    } catch (IncorrectInput e) {
-      view.messageNoItems(e.getMessage());
-    } catch (JRException e) {
-      if (ExceptionUtils.indexOfType(e.getCause(), PrinterAbortException.class) != -1) {
-        Tools.showPrintAbortedWarning(e, true);
-      } else {
-        Tools.showUnexpectedErrorWarning(e);
+      try {
+        throw e;
+      } catch (RuntimeException r) {
+        throw r.getCause();
       }
+    } catch (IncorrectInput i) {
+      getView().messageNoItems(e.getMessage());
+    } catch (UnsupportedOperationException u) {
+      getView().messageNotImplemented(exportType);
+    } catch (Throwable t) {
+      Tools.showUnexpectedErrorWarning(t);
     }
   }
 
   public void exportAccountingReport(ExportTypes exportType, int startBon, int endBon) {
     var view = getView();
     try {
-      model.exportAccountingReport(exportType, startBon, endBon);
+      model.exportAccountingReport(
+          exportType, startBon, endBon, (e) -> consumePdfException(e, exportType));
       view.back();
     } catch (UnsupportedOperationException e) {
       view.messageNotImplemented(exportType);
-    } catch (InvalidVATValueException e) {
-      view.messageNoItems(e.getMessage());
-    } catch (JRException e) {
-      if (ExceptionUtils.indexOfType(e.getCause(), PrinterAbortException.class) != -1) {
-        Tools.showPrintAbortedWarning(e, true);
-      } else {
-        Tools.showUnexpectedErrorWarning(e);
-      }
     }
   }
